Update a little of the refactor and have the system reach parity with the old!
