+++ src/plugin.rs
@@ -1,15 +1,21 @@
 use bevy::prelude::*;
+use bevy_tokio_tasks::TokioTasksPlugin;
 
-use crate::components::VideoFrame;
-use crate::systems::{camera_rotation, handle_video_frame, setup_physical_camera};
+use crate::bluetooth::{async_spawner, ZoetropeRotation};
+use crate::camera::VideoFrame;
+use crate::systems::{logical_camera_rotation, physical_camera_setup, update_zoetrope_image};
 
 pub struct ZoetropePlugin;
 
 impl Plugin for ZoetropePlugin {
     fn build(&self, app: &mut App) {
         app.insert_resource(VideoFrame(Handle::default()))
-            .add_startup_system(setup_physical_camera)
-            .add_system(handle_video_frame)
-            .add_system(camera_rotation);
+            .add_plugin(TokioTasksPlugin::default())
+            .insert_resource(ZoetropeRotation(0))
+            .add_startup_system(physical_camera_setup)
+            .add_startup_system(async_spawner)
+            .add_system(bevy::window::close_on_esc)
+            .add_system(update_zoetrope_image)
+            .add_system(logical_camera_rotation);
     }
 }
